requirements.txt
alembic==0.9.10
asn1crypto==0.24.0
astroid==1.6.5
cffi==1.11.5
click==6.7
cryptography==2.2.2
Flask==1.0.2
Flask-Markdown==0.3
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
py-bcrypt==0.4
pycparser==2.18
pylint==1.9.2
PyMySQL==0.9.0
python-dateutil==2.7.3
python-editor==1.0.3
python-slugify==1.2.5
six==1.11.0
SQLAlchemy==1.2.9
Unidecode==1.0.22
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
